TST: get the full test suite passing on pandas 3 / numpy 2 / Python 3.12
Library fixes (behaviour preserved from the pandas 0.22 era unless noted):
- data: keep datetimes at nanosecond resolution where they are stored or
  compared as int64 ns (HDF5 daily bars, FX, adjustments, resampling,
  benchmark sessions); fix silent-NaN lookups (Series.get with lists);
  select stock dividend columns by name instead of by position.
- pipeline: make generated hook methods concrete, allow null missing
  values in LabelArray categoricals, unify categoricals across chunks,
  sort output columns, and handle multiple-output factors without
  record arrays.
- finance/algorithm: copy-on-write safe daily returns buffer, accept
  interday capital changes keyed by UTC midnight, sort perf columns,
  compare naive restriction dates as UTC.
- assets: pass SQLAlchemy connections to to_sql; fix the exchange
  ambiguity check that old pandas made a no-op.
- utils: FinalMeta works without explicit bases, unzip handles empty
  input under PEP 479, session labels are nanosecond resolution.
- Cython: drop a typed loop variable in a closure that crashes under
  Cython 3.

Tests:
- Port the remaining tests to naive sessions, positional .iloc access,
  removed pandas APIs and SQLAlchemy 2; fix two helpers that became
  silent no-ops under copy-on-write.
- zipline.testing: honour keyword-only options of wrapped pandas
  assertion functions (check_names, check_freq were silently dropped).
- Distribute xdist workers by class (--dist=loadscope) and run doctests
  as a separate CI step.

2538 passed, 2 skipped (both skipped upstream); doctests pass.

+++ zipline/pipeline/loaders/events.py
@@ -211,7 +211,11 @@ class EventsLoader(PipelineLoader):
                      sids,
                      mask):
         def to_frame(array):
-            return pd.DataFrame(array, index=dates, columns=sids)
+            # Pass the dtype so pandas doesn't infer a string dtype (which
+            # would turn ``None`` missing values into NaN).
+            return pd.DataFrame(
+                array, index=dates, columns=sids, dtype=array.dtype,
+            )
 
         assert indexer.shape == (len(dates), len(sids))
